move nested-capable `progress_bar` method into the pipelines revert the original `progress_bar` method in pipeline_utils
